Rearranging License files
Renamed original LICENSE to GPL-2.0 and moved the license to LICENSE folder for easier management  of licenses.
Inspired by Trovalds/Linux kernel source way of licensing.

I am still learning about licenses and copyleft, so I have gone with a good point to start with, which is "GNU GPLv2" and not something extreme as "GNU GPLv3" or "GNU AGPLv3".
That is all for now on licenses and will update if needed or I became more greedy to the idea of copy left.